working refactor and tests
