+++ tests/utils.py
@@ -1,11 +1,8 @@
 import bpy
-
-# from .conftest import molecularnodes as mn
 import molecularnodes as mn
 
 import numpy as np
 import random
-# import pathlib
 
 from syrupy.extensions.amber import AmberSnapshotExtension
 
@@ -65,7 +62,7 @@ def sample_attribute(
     random.seed(seed)
     if error:
         attribute = mn.blender.mesh.named_attribute(
-            obj=object, naame=attribute, evaluate=evaluate
+            obj=object, name=attribute, evaluate=evaluate
         )
         length = len(attribute)
